+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Enable kube-proxy replacement mode in Cilium app.
+- Enable bootstrap mode for chart operator.
+
 ### Added
 
 - Add `vpc ID` to WC configmap on AWS.